requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7

# Translation
transformers>=4.40
sentencepiece>=0.1.99
sacremoses>=0.1.1

# Torch for RNN + Transformers
torch>=2.0

# Server + bot
fastapi>=0.110
uvicorn[standard]>=0.27
python-telegram-bot>=20.0
requests>=2.31
python-dotenv>=1.0

# Audio
pydub>=0.25
faster-whisper>=1.0
